[fidlc] 32 to 64 bits Ordinals: Step #6
fidlc now emits 64b ordinals in gen, causing all bindings to have to
read paths, one for ord (still 32b) and one for gen (now 64b)

Test: no behavior change
Bug: 7848

+++ zircon/tools/fidl/include/fidl/c_generator.h
@@ -63,9 +63,9 @@ class CGenerator {
   };
 
   struct NamedMethod {
-    uint32_t ordinal;
+    uint64_t ordinal;
     std::string ordinal_name;
-    uint32_t generated_ordinal;
+    uint64_t generated_ordinal;
     std::string generated_ordinal_name;
     std::string identifier;
     std::string c_name;
